+++ src/test/java/de/chrgroth/smartcron/model/Counter.java
@@ -1,6 +1,6 @@
 package de.chrgroth.smartcron.model;
 
-import java.util.Date;
+import java.time.LocalDateTime;
 
 import de.chrgroth.smartcron.api.Smartcron;
 
@@ -9,10 +9,10 @@ public abstract class Counter implements Smartcron {
     public int counter;
 
     @Override
-    public Date run() {
+    public LocalDateTime run() {
         counter++;
         return calc();
     }
 
-    protected abstract Date calc();
+    protected abstract LocalDateTime calc();
 }