+++ RELEASE-NOTES.md
@@ -1,13 +1,12 @@
 Release Notes
 =====================
 
-0.3.0 (in progress)
--------------------
+0.3.0
+-----
 - fixed bug in delay calculation
 - allowed smartcrons to be recoverable
 - enhanced smartcron metadata
 - switched to new Java 8 date time API
-- ...
 
 0.2.0
 -----